refactor NameEdit into separate file
