+++ packages/ui/src/Popup/Accounts/AccountView.tsx
@@ -1,54 +1,17 @@
 import { networkLinks } from '@polymathnetwork/extension-core/constants';
 import { IdentifiedAccount } from '@polymathnetwork/extension-core/types';
 import { recodeAddress } from '@polymathnetwork/extension-core/utils';
-import { SvgCheck, SvgDotsVertical, SvgPencilOutline, SvgWindowClose } from '@polymathnetwork/extension-ui/assets/images/icons';
+import { SvgCheck, SvgDotsVertical, SvgPencilOutline } from '@polymathnetwork/extension-ui/assets/images/icons';
 import BigNumber from 'bignumber.js';
 import React, { FC, useContext, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 
 import { AccountContext, AccountType, ActionContext, PolymeshContext } from '../../components';
 import { editAccount, setPolySelectedAccount } from '../../messaging';
-import { Box, ButtonSmall, ContextMenuTrigger, Flex, Icon, LabelWithCopy, Menu, MenuItem, Text, TextInput, TextOverflowEllipsis } from '../../ui';
+import { Box, ButtonSmall, ContextMenuTrigger, Flex, Icon, LabelWithCopy, Menu, MenuItem, Text, TextOverflowEllipsis } from '../../ui';
 import { formatters } from '../../util';
+import { NameEdit } from './NameEdit';
 import { AccountInfoGrid, AccountViewGrid, GridItem, UnassignedAccountHoverGrid } from './styles';
-
-type NameEditProps = {
-  name?: string;
-  newName?: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void
-  onSave: (e: React.MouseEvent<HTMLElement>) => Promise<void>
-  onCancel: (e: React.MouseEvent<HTMLElement>) => void;
-}
-
-const NameEdit: FC<NameEditProps> = ({ name, newName, onCancel, onChange, onSave }) => (<Box py='3px'>
-  <Flex alignItems='center'
-    flexDirection='row'>
-    <TextInput
-      defaultValue={name}
-      onChange={onChange}
-      tight
-      value={newName}
-    />
-    <Box ml='xs'>
-      <Icon
-        Asset={SvgCheck}
-        color='gray.2'
-        height={16}
-        onClick={onSave}
-        width={16}
-      />
-    </Box>
-    <Box ml='xs'>
-      <Icon
-        Asset={SvgWindowClose}
-        color='gray.2'
-        height={16}
-        onClick={onCancel}
-        width={16}
-      />
-    </Box>
-  </Flex>
-</Box>);
 
 export interface Props {
   account: IdentifiedAccount;
