+++ packages/ui/src/Popup/Accounts/NameEdit.tsx
@@ -1,0 +1,39 @@
+import { SvgCheck, SvgWindowClose } from '@polymathnetwork/extension-ui/assets/images/icons';
+import { Box, Flex, Icon, TextInput } from '@polymathnetwork/extension-ui/ui';
+import React from 'react';
+
+type NameEditProps = {
+  name?: string;
+  newName?: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onSave: (e: React.MouseEvent<HTMLElement>) => Promise<void>;
+  onCancel: (e: React.MouseEvent<HTMLElement>) => void;
+};
+
+export function NameEdit ({ name, newName, onCancel, onChange, onSave }: NameEditProps): JSX.Element {
+  return (
+    <Box py='3px'>
+      <Flex alignItems='center'
+        flexDirection='row'>
+        <TextInput defaultValue={name}
+          onChange={onChange}
+          tight
+          value={newName} />
+        <Box ml='xs'>
+          <Icon Asset={SvgCheck}
+            color='gray.2'
+            height={16}
+            onClick={onSave}
+            width={16} />
+        </Box>
+        <Box ml='xs'>
+          <Icon Asset={SvgWindowClose}
+            color='gray.2'
+            height={16}
+            onClick={onCancel}
+            width={16} />
+        </Box>
+      </Flex>
+    </Box>
+  );
+}
